% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affy_probeID_to_gene_symbols.R
\name{affy_probe_to_gene_symbol}
\alias{affy_probe_to_gene_symbol}
\title{use hgu133a.db or hgu133plus2.db to convert from affymetrix probe_id to human gene_symbol}
\usage{
affy_probe_to_gene_symbol(probe_id, GPL, sel.columns = "SYMBOL")
}
\arguments{
\item{genelist}{Gene list you want to convert}
}
\description{
Conversion between affy probe_ID to gene_symbol
}
\examples{
gene_convert_human_to_mouse("TP53")
}
\keyword{standardize}

