% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{assign_label_cex}
\alias{assign_label_cex}
\title{Adjust label cex based on its number}
\usage{
assign_label_cex(n.samples)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
assign_label_cex(100)
}
\keyword{cats}

