% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid_classify.R
\name{centroid_classify}
\alias{centroid_classify}
\title{Classify samples based on the feed in centroids}
\usage{
centroid_classify(x, cv_data, groups, method = "spearman")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{centroids}
\keyword{classify}

