% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_utils.R
\name{clean_iterNMF}
\alias{clean_iterNMF}
\title{Extract, cleanup, and merge iter_NMF results}
\usage{
clean_iterNMF(inmf_res, rawdata, cor.method = "spearman", cor.thresh = 0.9,
  ctriod.topN = 5000, ctroid.math = "mean", return.all = FALSE)
}
\arguments{
\item{rawdata}{gene count table for calculating centroids based on the clustering results}

\item{cor.method}{method for correlation calculation, default: spearman correlation}

\item{cor.thresh}{correlation threshold for merging subclusters}

\item{ctroid.math}{method for getting centroid, default: mean}

\item{iter_nmf}{vairous NMF res in a list format}

\item{ctroid.topN}{number of top Genes to extract for centroid comparison}
}
\description{
We will only get grouping info for now
}
\examples{
clean_iterNMF(inmf_res, rawdata)
}
\keyword{NMF}
\keyword{iterative}

