% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{clust_by_hier}
\alias{clust_by_hier}
\title{Clustering by hierarchical}
\usage{
clust_by_hier(data, K, takelog = FALSE, method = "ward.D")
}
\arguments{
\item{data}{integer data frame}
}
\description{
Clustering by hierarchical
}
\examples{
my.write.matrix(data)
}
\keyword{write.matrix}
