% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_utils.R
\name{clust_silhouette}
\alias{clust_silhouette}
\title{Calculate silhouette index between two data frame}
\usage{
clust_silhouette(clust1, clust2, method = "euclidean")
}
\arguments{
\item{clust1}{data in cluster1}

\item{clust2}{data in cluster2}

\item{method}{method for calculating distance metrics}
}
\description{
Calculate silhouette index between two data frame
}
\keyword{NMF}
\keyword{iterative}
\keyword{silhouette}
