% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_module_genelist.R
\name{comp_module_genelist}
\alias{comp_module_genelist}
\title{Loop through each of the modules and connect the adjacent ones which is only allow.gap away from each other}
\usage{
comp_module_genelist(final_comp_genelist, folderName0)
}
\arguments{
\item{final_summary}{summary table}

\item{final_genelist}{summary gene list}

\item{allow.gap}{allow how man gaps between each sub-modules}
}
\description{
Loop through each of the modules and connect the adjacent ones which is only allow.gap away from each other
}
\examples{
connect_gap_modules(final_summary, final_genelist)
}
\keyword{co-expression,}
\keyword{connectivity}
\keyword{network,}

