% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{compare_group}
\alias{compare_group}
\title{Compare two groups assignment results}
\usage{
compare_group(group1, group2, file.prefix = "Compare.group",
  title = "", plot = T, save.image = F, add.legend = T,
  label.size = 5, title.size = 36, angle = 45)
}
\arguments{
\item{group1}{vector of clustering 1}

\item{group2}{vector of clustering 2}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{compare}
