% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_groups.R
\name{compare_groups}
\alias{compare_groups}
\title{A Cat Function}
\usage{
compare_groups(group1, group2, file.prefix = "Compare.group", title = "",
  plot = T, save.image = F, add.legend = T, label.size = 5,
  title.size = 36)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

