% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{compare_groups_stats}
\alias{compare_groups_stats}
\title{Calculate statistics comparing two clustering results}
\usage{
compare_groups_stats(x, y, method = "rand", base = exp(1), beta = 1)
}
\arguments{
\item{x}{reference for cluster}

\item{y}{prediction for cluster}
}
\description{
Calculate statistics comparing two clustering results
}
\keyword{compare}
