% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{connect_gap_modules}
\alias{connect_gap_modules}
\title{Loop through each of the modules and connect the adjacent ones which is only allow.gap away from each other}
\usage{
connect_gap_modules(final_summary, final_genelist, allow.gap = 1)
}
\arguments{
\item{final_summary}{summary table}

\item{final_genelist}{summary gene list}

\item{allow.gap}{allow how man gaps between each sub-modules}
}
\description{
Loop through each of the modules and connect the adjacent ones which is only allow.gap away from each other
}
\examples{
connect_gap_modules(final_summary, final_genelist, allow.gap=1)
}
\keyword{co-expression,}
\keyword{connectivity}
\keyword{network,}

