% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{cor_mtest}
\alias{cor_mtest}
\title{load predefined gene list files for heatmap}
\usage{
cor_mtest(mat, ...)
}
\arguments{
\item{folder}{Where the files are}

\item{pattern}{file extensions for the ones you want to load}
}
\description{
Load predefined gene list files for heatmap (clustering) result
}
\examples{
cor_mtest("data", ".txt")
}
\keyword{lodad}

