% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{create.brewer.color}
\alias{create.brewer.color}
\title{Create brewer color}
\usage{
create.brewer.color(data, num = 8, name = "Set1")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
create.brewer.color()
}
\keyword{cats}
