% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.label.color.R
\name{create.label.color}
\alias{create.label.color}
\title{A Cat Function}
\usage{
create.label.color(data, color)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

