% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{create.manual.color}
\alias{create.manual.color}
\title{Manually specify coloring for provided groups}
\usage{
create.manual.color(data, group.color)
}
\arguments{
\item{data}{data}

\item{group.color}{colors for each factorial group}
}
\description{
Assign colors to data, if more unique data than provided colors, will impute the missing ones and fill them in
}
\examples{
create.manual.color(c(1,2,3,1,2,3,2,3), c("red", "blue", "green"))
}
\keyword{color}
\keyword{manual}
