% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_utils.R
\name{create_centroid}
\alias{create_centroid}
\title{Create centroid based on the grouping info}
\usage{
create_centroid(data, groups, scale = "none")
}
\arguments{
\item{data}{Input data}

\item{groups}{Groups for each samples}
}
\description{
This function allows you to express your love of cats.
}
\examples{
create_centroid(mtcars)
}
\keyword{centroid}
\keyword{reshape}
