% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{def_net_module2}
\alias{def_net_module2}
\title{Identify the the edge of the modules (blocks) within the net connectivity matrix}
\usage{
def_net_module2(core_sub_jaccard_dist, sub_jaccard_dist,
  min.connectivity = 0.2, min.size = 5, start_idx = 1,
  allow.gap = 1)
}
\arguments{
\item{core_sub_jaccard_dist}{preprocessed sub_jaccard_dist matrix (after applying filter, binarized the data)}

\item{sub_jaccard_dist}{zoomed-in small block size of jaccard_dist}

\item{min.connectivity}{threshold for minimum connectivity (set to 75\% quartile in each block)}

\item{min.size}{threshold for minimum block size}

\item{start_idx}{specifiy what is the index for the starting gene in the block}

\item{allow.gap}{allow how man gaps between each sub-modules}
}
\description{
Run for loops to detect the block
Detect edge genes in each block and their gene idx
Go back to the original 100x100 matrix and use the edge genes (idx) to define block
}
\examples{
def_net_modules2(core_sub_jaccard_dist, sub_jaccard_dist, min.connectivity=0.2, min.size=5)
}
\keyword{co-expression,}
\keyword{connectivity}
\keyword{network,}
