% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{def_net_modules}
\alias{def_net_modules}
\title{Identify the the edge of the modules (blocks) within the net connectivity matrix}
\usage{
def_net_modules(core_sub_jaccard_dist, sub_jaccard_dist,
  min.connectivity = 0.2, min.size = 5)
}
\arguments{
\item{data}{Input expression data}

\item{tao}{threshold for hard transformation}

\item{beta}{parameter for soft power transformation}

\item{num_features}{Number of top high connective genes to be used}
}
\description{
\preformatted{
1. Run for loops to detect the block
2. Detect edge genes in each block and their gene idx
3. Go back to the original 100x100 matrix and use the edge genes (idx) to define block
}
}
\examples{
def_net_modules(data, tao=0.5, beta=10, num_feature=0.01)
}
\keyword{co-expression,}
\keyword{connectivity}
\keyword{network,}

