% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{deseq_norm}
\alias{deseq_norm}
\title{Normalize data through DESeq size factor}
\usage{
deseq_norm(countdata)
}
\arguments{
\item{countdata}{integer data frame}
}
\description{
Normalize data through DESeq size factor
}
\examples{
deseq_norm(countdata)
}
\keyword{deseq}

