% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESeq2_Normalization.R
\name{DESeq_Normalization}
\alias{DESeq_Normalization}
\title{This function allows you to VST your data}
\usage{
DESeq_Normalization(countdata)
}
\arguments{
\item{countdata}{integer data frame}
}
\description{
This function allows you to VST your data
}
\examples{
vst(countdata)
}
\keyword{vst}

