% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{detect_genecnt_id}
\alias{detect_genecnt_id}
\title{Automatically convert genecnt into Entrez ID}
\usage{
detect_genecnt_id(data, sum.method = "mean")
}
\arguments{
\item{data}{Gene id list you want to convert}
}
\description{
Standize gene id into Entrez ID
}
\examples{
detect_genecnt_id("TP53")
}
\keyword{conversion}

