% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{detect_genecnt_platform}
\alias{detect_genecnt_platform}
\title{Check rownames from data and auto detect which platform it is from}
\usage{
detect_genecnt_platform(data, method = "mean")
}
\arguments{
\item{data}{Gene list you want to convert}
}
\description{
Basically we want to convert those affy metrix array data and just
leave normal gene_symbols intact
}
\examples{
detect_genecnt_platform("TP53")
}
\keyword{standardize}
