% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext_gene_function.R
\name{ext_gene_function}
\alias{ext_gene_function}
\title{Extract ensembl_id using gene symbol}
\usage{
ext_gene_function(genelist, ensembl, attributes = c("ensembl_gene_id",
  "hgnc_symbol", "description", "gene_biotype"))
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

