% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{extract_data_by_mad}
\alias{extract_data_by_mad}
\title{Extract data by MAD value}
\usage{
extract_data_by_mad(data, topN = 100, by = "row", type = "data")
}
\arguments{
\item{data}{Original gene expression data}

\item{topN}{How many genes from TopMAD list}
}
\description{
Extract data by MAD value
}
\examples{
extract_data_by_mad(mtcars, topN=10, type="data")
}
\keyword{MAD,}
\keyword{extract}
