% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{extract_data_by_math}
\alias{extract_data_by_math}
\title{Extract data by 'math' operation}
\usage{
extract_data_by_math(data, topN = 100, by = "row", type = "data",
  math = "mean")
}
\arguments{
\item{data}{Original gene expression data}

\item{topN}{How many genes from Top 'operation' list}
}
\description{
Extract data by 'math' operation
}
\examples{
extract_data_by_math(mtcars, topN=10, type="data", math="mean")
}
\keyword{data}
\keyword{extraction}
