% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_convert_human_to_mouse.R
\name{gene_convert_human_to_mouse}
\alias{gene_convert_human_to_mouse}
\title{use biomaRT to convert human gene to mouse gene}
\usage{
gene_convert_human_to_mouse(genelist)
}
\arguments{
\item{genelist}{Gene list you want to convert}
}
\description{
This function allows you to express your love of cats.
}
\examples{
gene_convert_human_to_mouse("TP53")
}
\keyword{standardize}

