% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{gene_convert_mouse_to_human}
\alias{gene_convert_mouse_to_human}
\title{use biomaRT to convert mouse gene to human gene}
\usage{
gene_convert_mouse_to_human(genelist)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
gene_convert_human_to_mouse("Trp53")
}
\keyword{standardize}
