% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ClaNC_group.R
\name{get_ClaNC_group}
\alias{get_ClaNC_group}
\title{A Cat Function}
\usage{
get_ClaNC_group(Build_out)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

