% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_utils.R
\name{getcentroid}
\alias{getcentroid}
\title{Get centroid from data frame}
\usage{
getcentroid(df, method = "mean")
}
\arguments{
\item{df}{gene count table}

\item{method}{mean or median for calculating centroid}
}
\description{
Get centroid from data frame
}
\keyword{NMF}
\keyword{centroid}
\keyword{iterative}
