% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_genelist.R
\name{load_genelist}
\alias{load_genelist}
\title{load predefined gene list files for heatmap}
\usage{
load_genelist(folder, pattern = ".txt")
}
\arguments{
\item{folder}{Where the files are}

\item{pattern}{file extensions for the ones you want to load}
}
\description{
Load predefined gene list files for heatmap (clustering) result
}
\examples{
load_genelist("data", ".txt")
}
\keyword{lodad}

