% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_finder.R
\name{module_finder}
\alias{module_finder}
\title{Identify the modules in Data within co-expression network}
\usage{
module_finder(data, p.value = 0.05, tao = 0.5, beta = 3,
  num_features = 100)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Specify the top percentile of correlation to be used
}
\examples{
cat_function()
}
\keyword{cats}

