% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{mutual_info_from_confmatrix}
\alias{mutual_info_from_confmatrix}
\title{Calculate Mutual information from confusion matrix}
\usage{
mutual_info_from_confmatrix(a, b, c, d)
}
\arguments{
\item{a}{True positive}

\item{b}{False negative}

\item{c}{False positive}

\item{d}{True negative}
}
\description{
Calculate Mutual information from confusion matrix
}
\keyword{information}
\keyword{mutual}

