% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{mutual_info_from_vector}
\alias{mutual_info_from_vector}
\title{calculate Mutual information from two vector}
\usage{
mutual_info_from_vector(x, y, base = exp(1))
}
\description{
first generate summary table, then use equation to calculate mutual info
}
