% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{my.write.matrix}
\alias{my.write.matrix}
\title{Own version of writing matrix output}
\usage{
my.write.matrix(x, file = "", sep = "\\t", col.names = T, append = F,
  row.names = F, justify = c("none", "left", "right"), pval = NULL,
  newline = "\\n", names = NULL)
}
\arguments{
\item{countdata}{integer data frame}
}
\description{
Own version of writing matrix output
}
\examples{
my.write.matrix(data)
}
\keyword{write.matrix}

