% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{myboxplot}
\alias{myboxplot}
\title{Custom version of boxplot}
\usage{
myboxplot(data, prefix = "Boxplot", method = "vioplot", title = "",
  ylim = c(0.6, 0.85), save.plot = F, axis.text.size = 14,
  axis.title.size = 16, legend.position = "right", vio.add.method = "dot")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
myboxplot()
}
\keyword{cats}

