% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfread.table.R
\name{myfread.table}
\alias{myfread.table}
\title{A Cat Function}
\usage{
myfread.table(filepath, check.platform = T, header = T, sep = "\\t",
  detect.file.ext = T)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to read data with fread (faster)
}
\examples{
cat_function()
}
\keyword{cats}

