% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_to_color.R
\name{name_to_color}
\alias{name_to_color}
\title{Convert name to color}
\usage{
name_to_color(name, split_pattern = "\\\\_", num_color = 1,
  ColScheme = c("naikai", "naikai2", "Set1", "Set2", "Set3", "Dark2",
  "Pastel1", "Pastel2", "Paired", "Accent"))
}
\arguments{
\item{name}{usually sample name, separate with "_"}
}
\description{
This function convert the sample_name in the data into different colors
}
\examples{
name_to_color()
}
\keyword{color}

