% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_utils.R
\name{nmf_extract_feature}
\alias{nmf_extract_feature}
\title{Extract features from NMF run results}
\usage{
nmf_extract_feature(res, rawdata = NULL, manual.num = 0,
  method = "default", math = "mad", FScutoff = 0.9)
}
\arguments{
\item{res}{NMF run results}

\item{rawdata}{Original expression matrix, if we want to rank the genes by 'MAD' value across samples}

\item{manual.num}{How many genes to select from each of the group}

\item{method}{'total', 'default', or 'rank'}

\item{math}{Default is 'mad': median absolute deviation}
}
\description{
X = W x H
W is the feature matrix, H is the sample matrix
After NMF run, use this function to select import features and assign groups for the two matrix
}
\examples{
nmf_extract_feature(nmf_res)
}
\keyword{NMF}
\keyword{feature}
