% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_utils.R
\name{nmf_extract_group}
\alias{nmf_extract_group}
\title{Extract groups (sample clustering) from NMF run results}
\usage{
nmf_extract_group(res, type = "consensus", matchConseOrder = F)
}
\arguments{
\item{res}{NMF run results}
}
\description{
X = W x H
W is the feature matrix, H is the sample matrix
After NMF run, use this function to select import features and assign groups for the two matrix
}
\examples{
nmf_extract_group(nmf_res)
}
\keyword{NMF}
\keyword{groups}
