% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_plot.R
\name{nmf_plot}
\alias{nmf_plot}
\title{Extract features from NMF run results}
\usage{
nmf_plot(res, type = "consensus", subsetRow = TRUE, save.image = F,
  hclustfun = "average", silorder = F, add_original_name = T)
}
\arguments{
\item{res}{NMF run results}

\item{mode}{What kind of plot? (result, feature, consensus)}
}
\description{
X = W x H
W is the feature matrix, H is the sample matrix
After NMF run, use this function to select import features and assign groups for the two matrix
}
\examples{
nmf_extract_feature(data)
}
\keyword{NMF}
\keyword{feature}

