% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_utils.R
\name{nmf_select_best_k}
\alias{nmf_select_best_k}
\title{Select the best k from multiple consensus NMF run}
\usage{
nmf_select_best_k(res, type = "consensus")
}
\arguments{
\item{res}{NMF run results}
}
\description{
X = W x H
W is the feature matrix, H is the sample matrix
After NMF run, use Cophenatic index to select the best k
highest value after k=2
}
\examples{
nmf_select_best_k(nmf_res)
}
\keyword{NMF}
\keyword{feature}

