% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_utils.R
\name{nmf_silhouette_plot}
\alias{nmf_silhouette_plot}
\title{Silhouette plot for NMF run}
\usage{
nmf_silhouette_plot(res, type = "consensus", silorder = F)
}
\arguments{
\item{res}{NMF run result}
}
\description{
This function allows you to express your love of cats.
}
\examples{
nmf_silhouette_plot(res)
}
\keyword{silhouette}
