% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_summary.R
\name{nmf_summary}
\alias{nmf_summary}
\title{Extract features from NMF run results}
\usage{
nmf_summary(res, class = NULL, save.data = F,
  filename = "nmf.summary.txt")
}
\arguments{
\item{res}{NMF run results}
}
\description{
X = W x H
W is the feature matrix, H is the sample matrix
After NMF run, use this function to select import features and assign groups for the two matrix
}
\examples{
nmf_summary(data)
}
\keyword{NMF}
\keyword{feature}

