% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{norm_factors}
\alias{norm_factors}
\title{Calculate DESeq normalization factor}
\usage{
norm_factors(data)
}
\arguments{
\item{data}{data to be transformed}
}
\description{
Calculate DESeq normalization factor
}
\examples{
norm_factors(mtcars)
}
\keyword{deseq,}
\keyword{normalization}
