% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_gene_expression}
\alias{plot_gene_expression}
\title{Plot expression for each gene}
\usage{
plot_gene_expression(data, gene, groups, type = "boxplot", data.return = f,
  ylim = c(0, 20), title = null, las = 1, cex.axis = 1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

