% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_modules.R
\name{plot_modules}
\alias{plot_modules}
\title{Loop through each of the modules and connect the adjacent ones which is only allow.gap away from each other}
\usage{
plot_modules(core_sub_jaccard_dist, final_genelist, allow.gap = 1)
}
\arguments{
\item{final_genelist}{summary gene list}

\item{allow.gap}{allow how man gaps between each sub-modules}

\item{final_summary}{summary table}
}
\description{
Loop through each of the modules and connect the adjacent ones which is only allow.gap away from each other
}
\examples{
connect_gap_modules(final_summary, final_genelist)
}
\keyword{co-expression,}
\keyword{connectivity}
\keyword{network,}

