% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_utils.R
\name{predict_from_centroid}
\alias{predict_from_centroid}
\title{Predict group assignment from centroids}
\usage{
predict_from_centroid(centroids, data, scale = "none", method = "spearman")
}
\arguments{
\item{centroids}{centroids for each group}

\item{data}{data to be classified}
}
\description{
Predict group assignment from centroids
}
\examples{
predict_from_centroid()
}
\keyword{centroid}

