% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{preprocessNetCon}
\alias{preprocessNetCon}
\title{Preprocess the connecitivty matrix based on mean connectivity and binary switch}
\usage{
preprocessNetCon(sub_jaccard_dist, rmv.filter = 0.5, binary.filter = 0.75,
  plot = F)
}
\arguments{
\item{data}{Input expression data}

\item{tao}{threshold for hard transformation}

\item{beta}{parameter for soft power transformation}

\item{num_features}{Number of top high connective genes to be used}
}
\description{
\preformatted{
1. First try to remove columns (rows) that have median expression 75% quantile
2. Convert the matrix into binary based on quantile connectivity'
}
}
\examples{
preprocessNetCon(sub_jaccard_dist, rmv.filter=0.5, binary.filter=0.75, plot=F)
}
\keyword{co-expression,}
\keyword{connectivity}
\keyword{network,}

