% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderfun.R
\name{reorderfun}
\alias{reorderfun}
\title{Reorder dendrogram based on mean expression for RNA-Seq}
\usage{
reorderfun(d, w)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
myHeatmap.3()
}
\keyword{cats}

