% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmv_constant_0.R
\name{rmv_constant_0}
\alias{rmv_constant_0}
\title{Remove genes that with too many constant 0 expression across samples}
\usage{
rmv_constant_0(data, by = "row", pct = 0.75, minimum = 0)
}
\arguments{
\item{data}{Input data set}

\item{by}{Default by "row", can change to "column"}
}
\description{
This function allows you to remove those data entries with constant variance
You can specify the percentage of 0s as the threshold
}
\examples{
rmv_constant_var(data, by="row")
Remove data entries with constant variance 
}
\keyword{rmv_constant_var}

