% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{rpm}
\alias{rpm}
\title{Transform the countdata into Reads per million reads}
\usage{
rpm(data, mapped_reads)
}
\arguments{
\item{data}{countdata}

\item{mapped_reads}{mapped reads in each library}
}
\description{
Transform the countdata into Reads per million reads
}
\examples{
rpm(data)
}
\keyword{RPM}

