% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{run.DESeq2}
\alias{run.DESeq2}
\title{Run DEseq2}
\usage{
run.DESeq2(sub.data, sub.expinfo, design, workers = 16)
}
\description{
This function allows you to run DESeq2 if provided required data
}
\examples{
run.DESeq2()
}
\keyword{DESeq2}

