% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ClaNC.R
\name{run_ClaNC}
\alias{run_ClaNC}
\title{A Cat Function}
\usage{
run_ClaNC(data, groups, ColSideColors, active.features = 20, est.num = 20,
  select.features = 10, file.prefix = "ClaNC", skip.est = F)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

