% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CoExpression.R
\name{run_CoExpression}
\alias{run_CoExpression}
\title{Preprocess the connecitivty matrix based on mean connectivity and binary switch 
1. First try to remove columns (rows) that have median expression 75% quantile
2. Convert the matrix into binary based on quantile connectivity'}
\usage{
run_CoExpression(expdata, thresh = "soft", tao = 0.7, beta = 10,
  method = "Pearson", folderName0 = "Module", avg_con_thresh = 0.00015,
  save.data = F)
}
\arguments{
\item{tao}{threshold for hard transformation}

\item{beta}{parameter for soft power transformation}

\item{data}{Input expression data}

\item{num_features}{Number of top high connective genes to be used}
}
\description{
Preprocess the connecitivty matrix based on mean connectivity and binary switch 
1. First try to remove columns (rows) that have median expression 75% quantile
2. Convert the matrix into binary based on quantile connectivity'
}
\examples{
preprocessNetCon(sub_jaccard_dist, rmv.filter=0.5, binary.filter=0.75, plot=F)
}
\keyword{co-expression,}
\keyword{connectivity}
\keyword{network,}

