% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_CV.R
\name{run_CV}
\alias{run_CV}
\title{A Cat Function}
\usage{
run_CV(cv_data, groups, nrun = 10, k.fold = 10, method = "spearman")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

