% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{run_DESeq2}
\alias{run_DESeq2}
\title{Run DEseq2}
\usage{
run_DESeq2(data, colData, design, workers = 8)
}
\description{
This function allows you to run DESeq2 if provided required data
}
\examples{
run_DESeq2()
}
\keyword{DESeq2}
