% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_utils.R
\name{run_tsne}
\alias{run_tsne}
\title{Use multi-core to run t-SNE}
\usage{
run_tsne(data, iter = 10, perplexity = 30, dims = 2, cores = 1)
}
\arguments{
\item{data}{expression data}

\item{iter}{how many iterations}

\item{perplexity}{perplexity}

\item{dims}{how many dimensions}

\item{cores}{how many cores to run t-SNE}
}
\description{
This function allows you to run t-SNE using multi-cores
}
\examples{
run_tsne(mtcars, iter=200, perplexity = 2, cores=4)
}
\keyword{t-SNE}

