% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sake-package.r
\docType{package}
\name{sake}
\alias{sake}
\alias{sake-package}
\title{sake.}
\description{
This tool can help you quickly navigate through the expression profile,
Using NMF method for unsupervised clustering.
}
