% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{scale_data}
\alias{scale_data}
\title{Scale data base on specify metrics}
\usage{
scale_data(x, scale = "row", na.rm = TRUE, method = "median")
}
\arguments{
\item{x}{data}

\item{scale}{By 'row' or 'column'}

\item{na.rm}{Remove NAs. Default is TRUE}

\item{method}{What kind of metrics? Default is 'median'. (Can be 'mean', 'median', 'mode', 'max', 'min', etc)}
}
\description{
This function allows you to scale the data based on your metrics of interest
}
\examples{
scale_data(mtcars)
}
\keyword{scale}
