% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{screenshot_plotly}
\alias{screenshot_plotly}
\title{Screen shot of plotly object}
\usage{
screenshot_plotly(x, file = "webshot.png", ...)
}
\arguments{
\item{x}{a plotly object}

\item{file}{output filename}
}
\description{
Screen shot of plotly object
}

