% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.n.pct.R
\name{select.n.pct}
\alias{select.n.pct}
\title{Select top n percentage of data}
\usage{
select.n.pct(data, n)
}
\arguments{
\item{data}{numerical data}

\item{n}{percentage to select}
}
\description{
Select top n pct from a vector
}
\examples{
select.n.pct(c(1,23,1,412,51,231,516,1,23,13,3,5,1), 10)
[1] 412 516
}
\keyword{n}
\keyword{percent}
\keyword{top}

