% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.top.n.R
\name{select.top.n}
\alias{select.top.n}
\title{Select top number from a vector}
\usage{
select.top.n(data, n, bottom = F)
}
\arguments{
\item{data}{numerical data}

\item{n}{top number}

\item{bottom}{whether to select from bottom?}
}
\description{
Select top # from a vector
}
\examples{
select.top.n(data, 100, bottom=F)
}
\keyword{select}

