% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{select_n_pct}
\alias{select_n_pct}
\title{Select top n percentage of data}
\usage{
select_n_pct(data, n)
}
\arguments{
\item{data}{numerical data}

\item{n}{percentage to select}
}
\description{
Select top n percentage of data
}
\examples{
select_n_pct(c(1,23,1,412,51,231,516,1,23,13,3,5,1), 10)
}
\keyword{n}
\keyword{percent}
\keyword{top}
