% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{select_top_n}
\alias{select_top_n}
\title{Select top number from a vector}
\usage{
select_top_n(data, n, bottom = F)
}
\arguments{
\item{data}{numerical data}

\item{n}{top number}

\item{bottom}{whether to select from bottom?}
}
\description{
Select top # from a vector
}
\examples{
select_top_n(mtcars, 100, bottom=F)
}
\keyword{select}
