% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleCap.R
\name{simpleCap}
\alias{simpleCap}
\title{Capitalize the first letter}
\usage{
simpleCap(x)
}
\arguments{
\item{genelist}{Gene list you want to convert}
}
\description{
This function allows you to express your love of cats.
}
\examples{
simpleCap("meat")
}
\keyword{standardize}

