% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_module2.R
\name{summarize_module2}
\alias{summarize_module2}
\title{Identify the the edge of the modules (blocks) within the net connectivity matrix
Run for loops to detect the block
Detect edge genes in each block and their gene idx
Go back to the original 100x100 matrix and use the edge genes (idx) to define block}
\usage{
summarize_module2(core_sub_jaccard_dist, old, current, blk_cnt, genenames,
  original_names, res_summary, res_genelist, size, min.size = 5, start_idx)
}
\arguments{
\item{core_sub_jaccard_dist}{}
}
\description{
Extract groups (sample clustering) from NMF run results 
X = W x H
W is the feature matrix, H is the sample matrix
After NMF run, use this function to select import features and assign groups for the two matrix
}
\examples{
summarize_module(core_sub_jaccard_dist, sub_jaccard_dist, min.connectivity=0.2, min.size=5)
}
\keyword{NMF}
\keyword{feature}

