% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_genes.R
\name{top_genes}
\alias{top_genes}
\title{Select top n genes}
\usage{
top_genes(data, n = 20, whole = F, name = "Exp1")
}
\arguments{
\item{data}{numerical data}

\item{n}{number to select}

\item{whole}{whether to return the whole sets}
}
\description{
Select top n genes
}
\examples{
top_genes(data)
}
\keyword{n}
\keyword{percent}
\keyword{top}

