% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UQ.R
\name{UQ}
\alias{UQ}
\title{Transform the countdata by upper quantile value from each sample}
\usage{
UQ(data, remove.zero = T)
}
\arguments{
\item{data}{countdata}
}
\description{
Transform the countdata using the quantile value from each sample
}
\examples{
RPM()
}
\keyword{UQ,}
\keyword{normalization}
\keyword{quantile,}

