% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Usoskin}
\alias{Usoskin}
\title{Data from Usoskin et al}
\format{A data frame with 25333 genes and 864 samples:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.nature.com/neuro/journal/v18/n1/abs/nn.3881.html}
}
\usage{
Usoskin
}
\description{
Title: 'Unbiased classification of sensory neuron types
by large-scale single-cell RNA sequencing'
}
\details{
Abstract:
The primary sensory system requires the integrated function of multiple cell types,
although its full complexity remains unclear. We used comprehensive transcriptome
analysis of 622 single mouse neurons to classify them in an unbiased manner, independent
of any a priori knowledge of sensory subtypes. Our results reveal eleven types: three
distinct low-threshold mechanoreceptive neurons, two proprioceptive, and six principal
types of thermosensitive, itch sensitive, type C low-threshold mechanosensitive and
nociceptive neurons with markedly different molecular and operational properties.
Confirming previously anticipated major neuronal types, our results also classify and
provide markers for new, functionally distinct subtypes. For example, our results suggest
that itching during inflammatory skin diseases such as atopic dermatitis is linked to a
distinct itch-generating type. We demonstrate single-cell RNA-seq as an effective strategy
for dissecting sensory responsive cells into distinct neuronal types. The resulting
catalog illustrates the diversity of sensory types and the cellular complexity underlying
somatic sensation.
}
\keyword{datasets}

