% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vst.R
\name{vst}
\alias{vst}
\title{This function allows you to VST your data}
\usage{
vst(countdata)
}
\arguments{
\item{countdata}{integer data frame}
}
\description{
This function allows you to VST your data
}
\examples{
vst(countdata)
}
\keyword{vst}

