% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_utils.R
\name{weighted.silhouette}
\alias{weighted.silhouette}
\title{calculate weighted silhouette_width as threshold for running iter_NMF}
\usage{
weighted.silhouette(nmf_groups)
}
\arguments{
\item{nmf_groups}{grouping infor after NMF run}
}
\description{
calculate weighted silhouette_width as threshold for running iter_NMF
}
