% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{wgcna_ext_hubgenes}
\alias{wgcna_ext_hubgenes}
\title{Extract hub genes in each module by using WGCNA}
\usage{
wgcna_ext_hubgenes(datExpr, MM.cutoff = 0.7, math.cutoff.quantile = 3,
  math = "mad", top.math = 1500, plot = F, save.data = F,
  prefix = "MM_vs_MAD")
}
\arguments{
\item{datExpr}{gene count data}

\item{MM.cutoff}{module membership cutoff}

\item{math.cutoff.quantile}{which quantile to use as min cutoff}

\item{math}{default is 'mad': can be 'mean', 'median', 'iqr'}
}
\description{
Weighted gene correlation network analysis
}
\examples{
wgcna_ext_hubgenes(datExpr)
}
\keyword{co-expression,}
\keyword{network}
\keyword{wgcna,}
