% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_extract_hubgenes.R
\name{wgcna_extract_hubgenes}
\alias{wgcna_extract_hubgenes}
\title{Extract features from NMF run results}
\usage{
wgcna_extract_hubgenes(datExpr, MM.cutoff = 0.7, math.cutoff.quantile = 3,
  math = "mad", top.math = 1500, plot = F, save.data = F,
  prefix = "MM_vs_MAD")
}
\arguments{
\item{math}{Default is 'mad': median absolute deviation}

\item{res}{NMF run results}

\item{rawdata}{Original expression matrix, if we want to rank the genes by 'MAD' value across samples}

\item{manual.num}{How many genes to select from each of the group}

\item{method}{'total', 'default', or 'rank'}
}
\description{
X = W x H
W is the feature matrix, H is the sample matrix
After NMF run, use this function to select import features and assign groups for the two matrix
}
\examples{
nmf_extract_feature(nmf_res)
}
\keyword{NMF}
\keyword{feature}

