% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{comp_module_genelist}
\alias{comp_module_genelist}
\title{Compute module gene list}
\usage{
comp_module_genelist(final_comp_genelist, folderName0)
}
\arguments{
\item{final_comp_genelist}{gene list}

\item{folderName0}{file folder}
}
\description{
Compute module gene list
}
\examples{
comp_module_genelist(final_comp_genelist, folderName0="folder")
}
\keyword{co-expression,}
\keyword{connectivity}
\keyword{network,}
