% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_utils.R
\name{ext_gene_function}
\alias{ext_gene_function}
\title{Extract ensembl_id using gene symbol}
\usage{
ext_gene_function(genelist, ensembl, attributes = c("ensembl_gene_id",
  "hgnc_symbol", "description", "gene_biotype"))
}
\arguments{
\item{genelist}{}

\item{ensembl}{}
}
\description{
Can be extened by providing the "columan name" that we want to match in snpinfo #
}
\examples{
ext_gene_function()
}
\keyword{annotate}
\keyword{ensembl,}

