% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data_by_mad.R
\name{extract_data_by_mad}
\alias{extract_data_by_mad}
\title{Extract data by MAD value}
\usage{
extract_data_by_mad(data, topN, by = "row", type = "data")
}
\arguments{
\item{data}{Original gene expression data}

\item{topN}{How many genes from TopMAD list}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

