% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanc_utils.R
\name{get_ClaNC_group}
\alias{get_ClaNC_group}
\title{Get ClaNA Group info}
\usage{
get_ClaNC_group(Build_out)
}
\arguments{
\item{Build_out}{}
}
\description{
Extract which group each point belongs to
}
\examples{
get_ClaNC_group()
}
\keyword{ClaNC}
