% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_utils.R
\name{module_finder}
\alias{module_finder}
\title{Identify the modules in Data within co-expression network}
\usage{
module_finder(data, p.value = 0.05, tao = 0.5, beta = 3,
  num_features = 100)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
Identify the modules in Data within co-expression network
}
\examples{
module_finder(mtcars)
}
\keyword{co-expression}
\keyword{module,}
\keyword{network,}

