% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{myfread.table}
\alias{myfread.table}
\title{Use data.table for faster read.table}
\usage{
myfread.table(filepath, check.platform = T, header = T, sep = "\\t",
  detect.file.ext = T)
}
\arguments{
\item{filepath}{input file path}
}
\description{
Use data.table for faster read.table
}
\examples{
myfread.table(system.file('sake', package='sake'))
}
\keyword{data}
\keyword{read}
