% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetConnectivity.R
\name{NetConnectivity}
\alias{NetConnectivity}
\title{Identify the informative genes in Data with top high connectivity in the co-expression network}
\usage{
NetConnectivity(data, thresh = "soft", tao = 0.7, beta = 1,
  method = "pearson", diag.zero = T)
}
\arguments{
\item{data}{Input expression data}

\item{tao}{threshold for hard transformation}

\item{beta}{parameter for soft power transformation}

\item{num_features}{Number of top high connective genes to be used}
}
\description{
Code is adapted and rewrote based on Wange et al, BMC Bioinformatics, 2014 
'Improving the sensitivity of sample clustering by leveraging gene co-expression networks in variable selection'
}
\examples{
NetConnectivity(data, tao=0.5, beta=10, num_feature=0.01)
}
\keyword{co-expression,}
\keyword{connectivity}
\keyword{network,}

