% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_utils.R
\name{nmf_estim_plot}
\alias{nmf_estim_plot}
\title{Generate individual plot for estimating K}
\usage{
nmf_estim_plot(estim.r)
}
\arguments{
\item{res}{NMF run result}
}
\description{
This function allows you to express your love of cats.
}
\examples{
nmf_estim_plot(estim.r)
}
\keyword{silhouette}
