% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_factors.R
\name{norm_factors}
\alias{norm_factors}
\title{A Cat Function}
\usage{
norm_factors(mat)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

