% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_utils.R
\name{parse_tsne_res}
\alias{parse_tsne_res}
\title{Extract and parse tsne results}
\usage{
parse_tsne_res(tsne_out)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
tsne_out <- run_tsne(mtcar, perplexity=3)
parse_tsne_res(tsne_out)
}
\keyword{cats}
