% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_from_centroid.R
\name{predict_from_centroid}
\alias{predict_from_centroid}
\title{A Cat Function}
\usage{
predict_from_centroid(centroids, data, scale = "none", method = "spearman")
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

