% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{reorderfun}
\alias{reorderfun}
\title{Reorder dendrogram based on mean expression for RNA-Seq}
\usage{
reorderfun(d, w)
}
\arguments{
\item{d}{dendrogram}

\item{w}{dendrogram}
}
\description{
Reorder dendrogram based on mean expression for RNA-Seq
}
\examples{
reorderfun()
}
\keyword{reorder}
