% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{rmv_constant_0}
\alias{rmv_constant_0}
\title{Remove genes that with too many constant 0 expression across samples}
\usage{
rmv_constant_0(data, by = "row", pct = 0.75, minimum = 0,
  verbose = FALSE)
}
\arguments{
\item{data}{Input data set}

\item{by}{Default by "row", can change to "column"}
}
\description{
Allow you to remove those data entries with constant variance
You can specify the percentage of 0s as the threshold
}
\examples{
rmv_constant_0(data, by="row", pct=0.75)
}
\keyword{rmv_constant_var}

