% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPM.R
\name{RPM}
\alias{RPM}
\title{Transform the countdata into Reads per million reads}
\usage{
RPM(data, mapped_reads)
}
\arguments{
\item{data}{countdata}

\item{mapped_reads}{mapped reads in each library}
}
\description{
Transform the countdata into Reads per million reads
}
\examples{
RPM()
}
\keyword{RPM}

