% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_utils.R
\name{run_CV}
\alias{run_CV}
\title{Run Cross-Validation}
\usage{
run_CV(cv_data, groups, nrun = 10, k.fold = 10, method = "spearman")
}
\arguments{
\item{cv_data}{}
}
\description{
Run Cross-Validation
}
\examples{
run_CV(cv_data)
}
\keyword{cross-validation}
