% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{simpleCap}
\alias{simpleCap}
\title{Convert letter to capitalization}
\usage{
simpleCap(x)
}
\arguments{
\item{x}{character}
}
\description{
Convert letter to capitalization
}
\examples{
simpleCap("meat is good")
}
\keyword{standardize}

