% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{uq}
\alias{uq}
\title{Transform the countdata by upper quantile value from each sample}
\usage{
uq(data, remove.zero = T)
}
\arguments{
\item{data}{countdata}
}
\description{
Upper quartile normalization, add option to remove all zeros in the sample first
}
\examples{
uq(mtcars)
}
\keyword{UQ,}
\keyword{normalization}
\keyword{quantile,}

