% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{vst}
\alias{vst}
\title{Use vst (fast) in DESeq2}
\usage{
vst(countdata, fast = TRUE)
}
\arguments{
\item{countdata}{integer data frame}
}
\description{
Use vst (fast) in DESeq2
}
\examples{
vst(countdata)
}
\keyword{deseq2}
\keyword{vst,}

